locals/jobs-scraper/requirements.txt
python-jobspy
langchain
langchain-openai
langchain-community
